data/requirements.txt
pandas>=2.2.3
psycopg2_binary>=2.9.10
python-dotenv>=1.0.1
python_dateutil>=2.9.0.post0
pyvirtualdisplay>=3.0
Requests>=2.32.3
selenium>=4.27.1
webdriver_manager>=4.0.2
